Make goose build and pass the test suite again.
Expose SetupHTTP, since that seems to be how the service is meant to be used.
Change the error classes to not use global state, but instead grab the values
from the new '*Nova' attribute. The static errors won't have one, but that
didn't break any tests. Either the testing is incomplete, or the point where
we rewrap any errors into an actual error is handling it correctly.
Change lbox.check so that we know 'go build ./...' is happy, so we will
be less likely to run into this in the future.
Note that none of the 'local_tests' actually used the nova double, so
having it 'working' is a bit of a misnomer. It just builds, it wasn't
being used anyway.