Add CompatLogger interface, loggo impl
We change several functions to take a new
CompatLogger instead of *log.Logger, and
provide a way of upgrading the interface
to a loggo.Logger.

Technically this is an API break, because
the signature of several exported functions
have changed. The CompatLogger interface is
implemented by *log.Logger, so if you're
just calling the functions and not taking
the function values, you'll be fine.