avoid arbitrary memory use in upload
When uploading a file, the code was reading the entire
body into memory before making the request.
This caused out-of-memory errors when
uploading large files.

Now, if you provide a ReadSeeker, it will seek to
the start of the file rather than reading the whole
thing into memory.

It's tempting to change the API so that it requires
a ReadSeeker rather than a Reader, but that would break
the goose.v2 compatibility guarantee.