Port V2 API calls
Adding the ability to create, get, list and delete a series of ports
for the neutron API.

The following code duplicates most of the existing work found with in
the neutron networking code, but adapts it for Ports. The API
documentation for the Ports API can be found[1].

This is required to provide spaces and multiple nic support to Juju. See
work in progress PR[2]

1. https://docs.openstack.org/api-ref/network/v2/index.html?expanded=create-port-detail#ports
2. https://github.com/juju/juju/pull/11188