Add ports http service testing
The following adds tests to ensure that the http server works as
expected. The feels a lot like testing the tests at which point you have
to ask if this is overkill?

I feel in the future it would be better served to write mocks for the
API and not have so much http testing tests. One has to realise that
you're actually testing the tests and not the integration with the
openstack API.