Add List filtering
Adding list filtering is an expectation when using the goose library and
it has it in other places. The server code was setup for this, but the
library code wasn't able to pass that in. Adding the filtering was
really simple by just passing in the generic filter option.

Additionally fixing typo and ensuring we're exposing everything we need
when requesting the port.