Omit empty on FixedIPsV2
The following ensures that we don't have to supply a fixed ip address if
you're not likely to have one. Also we invert that statement as well, as
you may have a IPAddress but not a subnet ID, so in both scenarios we
omitempty.

This should help with trying to create a port with in Juju, but we don't
have a subnet_id.