Neutron model test access
In order to replicate how openstack performs certain duties at the
service layer, we need access to the underlying model to be able to
perform the same duties. This then allows us to exercise the code in
Juju.

In an ideal world this wouldn't need to be done, as Juju would have unit
tested at the right level instead of using blackbox testing at every
level.