Port tags
In order to identify ports once created and to be able to filter
them using list ports, it's ideal to place tags on each port.

The code updates the port struct to add the tag and then goes
through the test harness to expect the tag field in the tests.