Add os-interfaces to nova
The following adds the /servers/{server_id}/os-interface endpoint
for listing all ports associated with a server.

Added subsequent tests to the stubs to ensure that we test the getting
the expected interfaces back from the stub server.

The documentation for this is here:
https://docs.openstack.org/api-ref/compute/?expanded=list-port-interfaces-detail

The is required to fulfill the need from Juju to be able to get all the
ports of a given server, to then be able to delete them when destroying
a controller.