Expose http headers
The following changes expose the http headers to we as a library user
can alter them if required. This happens to be the case for juju, where
we want to ensure that some headers are not set.

This is a halfway house as the real fix is to create a factor method for
handling the different client constructors. Then you can pass in a
header method to the factory to then create them. Unfortunately we're
late in the cycle and doing that at this late stage is _very_ risky. So
let's take the more moderate approach and come back to this for V3 of
goose.