HTTP: Allow passing http client
Passing in the http client, allows us to ensure the same client is used
throughout the lifecycle of a goose query. The previous implementations
either used the default http client in the stdlib or it created a new
http.client depending on circumstances of how you create the goose http
client.

The code in question tries to be as backwards compatible as possible so
we don't need to rev the library. There are some deprecated ways to
create a goose http client that are no longer in use internally, but are
still made available.