Logging: Default logger
Remove the loggo dependency from the non-test code base. You don't need
a loggo dependency directly for logging, instead you just need a Debugf
method. The fact that a direct dependency on this was required makes it
very difficult to iterate the loggo logger in the new modern world of go
mod.

Remove the comment about just depending on loggo, I really think that's
the wrong thing to do.